feat: restrict edit and delete rooms functionality to allowed users

+++ studybud/base/views.py
@@ -3,6 +3,7 @@ from django.contrib import messages
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
@@ -67,6 +68,8 @@ def create_room(request):
 @login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     form = RoomForm(instance=room)
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -80,6 +83,8 @@ def update_room(request, pk):
 @login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     if request.method == "POST":
         room.delete()
         return redirect("home")
